backend/requirements.txt
alembic==1.13.3
bcrypt==4.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cmake==3.25.2
dlib==19.24.2
dnspython==2.7.0
eventlet==0.37.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
h5py==3.12.1
idna==3.10
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.5
MarkupSafe==3.0.1
numpy==1.23.5
opencv-python==4.8.0.74
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
requests==2.32.3
scikit-learn==1.3.0
scipy==1.14.1
simple-websocket==1.1.0
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
transformers
midas
torch
torchvision